module api: proper documentation for the ibuf wrapper
We have extended module api as part of #5384 but there were
some complains that this part ofapi is not self-contained
and it's impossible to understand what's going on there
without looking into small/ibuf.[hc] sources.

Now we try to document the essential part of ibuf structure
so API should makes much more sense now.

Follow-up to #5384