test: update result files considering ujit hashf
LuaJIT has as different hash function as it has an additional bloom
filter for string hashes. As far as a result file should strictly
duplicate word to word results of console's output its order depends on
a hash function.

Also LuaJIT crops error messages opposed to LuaVela. The result file for
some test is different but the entity is the same.

This patch updates result files for corresponding tests as they actually
pass.

Relates to #5101