DO NOT MERGE: tarantool -e test fixup
Background: the test fails on waiting the process termination, but
surprisingly only on GitHub hosted runners.

My guess is that the process becomes zombie, but we don't collect it,
because on the busy loop. AFAIK, libev processes SIGCHLD and executes
wait(). So we should give it a change to do so.

Aside of all those things, a busy loop is not good in general. It is
better to pass execution to other fibers if there is nothing to do. This
way the waiting will not eat 100% CPU time.