memtx: implement api for memory allocator selection
Added an allocator option in the.cfg window for {}, which provides
user the ability to select a memory allocator for memtx.
The available option values is "system" and "small". The default value
is "small".

Closes #5419

@TarantoolBot document
Title: Add new 'allocator' option to box.cfg{}
Add new 'allocator' option to box.cfg{} which allows to
select the appropriate allocator for memtx tuples if necessary.
Possible values are "system" for malloc allocator and
"small" for default small allocator.