Implemented on_shutdown API
Implemented on_shutdown API, which allows to register functions
that will be called when the tarantool stopped. Functions will
be called in the reverse order they are registered, so the developer
has two options for API using:
- If on_shutdown module function completes quickly, module developer
  simply registers it using on_shutdown API.
- Otherwise module developer registers two functions in the following order,
  the first one waits for module works completion, the second one starts module
  termination process.

@TarantoolBot document
Title: Implement on_shutdown API
Implemented on_shutdown API, which allows to register functions
that will be called when the tarantool stopped. Functions will
be called in the reverse order they are registered, so the developer has
two options for API using:
- If on_shutdown module function completes quickly, module developer
  simply registers it using on_shutdown API.
- Otherwise module developer registers two functions in the following order,
  the first one waits for module works completion, the second one starts module
  termination process.

Closes #5723