Implemented test for on_shutdown module API
Implemented simple module, that checks on_shutdown API
functionality. Module first registers function that waits
for the module works completion. This function must be
registered first, because the functions are run in the reverse
order of their registration. Then module registers the first
on_shutdown function, which stops the module, then replaces it
with a new one.