Fixed force_recovery test.
Test checks the possibility of recovery with force_recovery option.
For these purpose, snapshot is damaged in test and possibility of
recovery is checked. In previous version snapshot damaged during
server worked, which sometimes made it impossible to recover.
Since such scenario is not considered, now snapshot corruption
occurs only when server is turned off.

Follow-up #5422