vinyl: remove vylog newer than snap in casual recovery
As a follow-up to the previous patch, let's check also emptiness of the
vylog being removed. If it contains no other entries except SNAPSHOT
(meaning that vylog is in fact empty), we can safely remove it without
hesitation. So it is OK to remove it even during casual recovery
process. However, if it contains rows after SNAPSHOT marker, removal of
vylog may cause data loss. In this case we still can remove it only
in force_recovery mode.

Follow-up #5823